requirements.txt
git+https://github.com/fra31/auto-attack.git

einops==0.4.1

folder==0.0.1

h5py==3.7.0

matplotlib==3.5.3

numpy==1.23.3

pandas==2.2.2

Pillow==9.2.0

pycountry==22.3.5

scipy==1.9.1

setuptools==68.0.0

torch==2.0.1

torchvision==0.15.2

tqdm==4.64.1

streamlit==1.24.1

opencv-python==4.8.0.74

clip===1.0